% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{AddGaussianNoise}
\alias{AddGaussianNoise}
\title{Additive Gaussian Noise}
\usage{
AddGaussianNoise(y, prob, mu, sigma)
}
\arguments{
\item{prob}{The probability of success of the binomial distribution. A number between 0 and 1.}

\item{mu}{The mean of the normal distribution.}

\item{sigma}{The standard deviation of the normal distribution. A real number.}

\item{Y}{The time-series (as numeric) to be contaminated.}
}
\value{
The y matrix with additive noise.
}
\description{
Generates a matrix _C_ from an initial matrix _Y_ and a matrix _Nu_ of the same 
dimension. _Nu_ values follow an _N_(_mu_, _sigma_^2) distribution. The _prop_% of 
the elements of the array _Y_ are selected randomly. If Y_ij_ is selected 
then it is replaced by _Y_ij_ + _Nu_ij_. This procedure is called additive 
contamination of _prob_%.
}
