% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{AmpSca}
\alias{AmpSca}
\title{Amplitude Scaling variation imposed}
\usage{
AmpSca(S, beta = 1)
}
\arguments{
\item{S}{A time series with real values.}

\item{beta}{A constant that is multiplied to all point in the time series.}
}
\value{
The input time series with the amplitude scaling variation imposed.
}
\description{
A time series distortion that modify the amplitude scale of a time series _S_.
}
\details{
$S1(t) = beta * S(t)$
}
