% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{AmpShift}
\alias{AmpShift}
\title{Amplitude Shift variation imposed}
\usage{
AmpShift(S, beta = 1)
}
\arguments{
\item{S}{A time series with real values.}

\item{beta}{A constant that is added to all points in the time series.}
}
\value{
The input time series with the amplitude shift variation imposed.
}
\description{
A time series distortion that shift the amplitude range of a time series _S_.
}
\details{
$S1(t) = beta + S(t)$
}
