% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{Chebyshev.Measure}
\alias{Chebyshev.Measure}
\title{CHEBYSHEV MEASURE}
\usage{
Chebyshev.Measure(s1, s2)
}
\arguments{
\item{s1}{a numeric vector containing the first time series.}

\item{s2}{a numeric vector containing the second time series.}
}
\value{
a positive real number. The computed distance between the pair of series.
}
\description{
Computes the Chebyshev distance using the 'chebyshev' function from the 'philentropy' package by setting the testNA parameter as False.
}
\references{
for more details read the 'philentropy' package documentation.
}
