% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{CHOU}
\alias{CHOU}
\title{CHOUAKRIA INDEX}
\usage{
CHOU(S1, S2, k = 3.1)
}
\arguments{
\item{S1}{a time series of length n with real values.}

\item{S2}{a time series of length n with real values.}

\item{k}{a equal to 3.1 by default.}
}
\value{
D a positive real number plus 0; where 0 means max similarity.

CORT temporal correlation between S1 and S2.

dE point-to-point euclidean distance between S1 and S2.
}
\description{
Compute D index of Chouakria and Nagabhushan similarity index between two same length time series S1 and S2.
}
\references{
for more details read Chouakria and Nagabhushan 2007.
}
