% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{CORT}
\alias{CORT}
\title{TEMPORAL CORRELATION}
\usage{
CORT(x, y)
}
\arguments{
\item{S1}{a time series of length n with real values.}

\item{S2}{a time series of length n with real values.}

\item{k}{a equal to 3.1 by default.}
}
\value{
CORT temporal correlation between S1 and S2.
}
\description{
Compute the Temporal Correlation of Chouakria and Nagabhushan between two same length time series S1 and S2.
}
\references{
for more details read  Chouakria Douzal 2003.
}
