% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{DTW.DIST}
\alias{DTW.DIST}
\title{DYNAMIC TIME WARPING WITH NA.OMIT}
\usage{
DTW.DIST(x, y)
}
\arguments{
\item{x}{a time series with real values.}

\item{y}{a time series with real values and not necessarily the same length as x.}
}
\value{
a positive real number result of dynamic time warping distance between x and y using dtw(x,y)$distance function.
}
\description{
Compute dynamic time warping distance between two time series x and y using the "dtw" library.
}
