% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{dtw_mse}
\alias{dtw_mse}
\title{DYNAMIC TIME WARPING MSE
Computes the Dynamic Time Warping mapped path between a pair of numeric time series using the 'dtw' function from the 'dtw' package.
and calculate the MSE of the mapped path.
We by setting the 'dtw' function by using a  Sakoe-Chiba windowing constraint with a window equal to the 10% of the time series length.}
\usage{
dtw_mse(S1, S2)
}
\arguments{
\item{s1}{a numeric vector containing the first time series.}

\item{s2}{a numeric vector containing the second time series.}
}
\value{
the computed distance between the pair of series.
}
\description{
DYNAMIC TIME WARPING MSE
Computes the Dynamic Time Warping mapped path between a pair of numeric time series using the 'dtw' function from the 'dtw' package.
and calculate the MSE of the mapped path.
We by setting the 'dtw' function by using a  Sakoe-Chiba windowing constraint with a window equal to the 10% of the time series length.
}
\references{
for more details read the 'dtw' package documentation.
}
