% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{EDR.Measure}
\alias{EDR.Measure}
\title{EDITH DISTANCE FOR REAL SEQUENCE
Computes the Edit Distance for Real Sequence between a pair of numeric time series using the 'EDRDistance' function from the 'TSdist' package.
We by setting the 'sigma' parameter as the 10% of the time series length, and the 'epsilon' parameter equal to 0.1.}
\usage{
EDR.Measure(s1, s2)
}
\arguments{
\item{s1}{a numeric vector containing the first time series.}

\item{s2}{a numeric vector containing the second time series.}

\item{epsilon}{a positive threshold value that defines the distance.}

\item{sima}{a Sakoe-Chiba windowing constraint can be added by specifying a positive integer representing the window size.}
}
\value{
the computed distance between the pair of series.
}
\description{
EDITH DISTANCE FOR REAL SEQUENCE
Computes the Edit Distance for Real Sequence between a pair of numeric time series using the 'EDRDistance' function from the 'TSdist' package.
We by setting the 'sigma' parameter as the 10% of the time series length, and the 'epsilon' parameter equal to 0.1.
}
\references{
for more details read the 'TSdist' package documentation.
}
