% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{ERP.Measure}
\alias{ERP.Measure}
\title{EDITH DISTANCE WITH REAL PENALTY 
Computes the Edit Distance with Real Penalty between a pair of numeric time series using the 'ERPDistance' function from the 'TSdist' package.
We by setting the 'sigma' parameter as the 10% of the time series length, and the 'g' parameter equal to 0.1.}
\usage{
ERP.Measure(s1, s2)
}
\arguments{
\item{s1}{a numeric vector containing the first time series.}

\item{s2}{a numeric vector containing the second time series.}

\item{g}{the reference value used to penalize gaps.}

\item{sima}{a Sakoe-Chiba windowing constraint can be added by specifying a positive integer representing the window size.}
}
\value{
the computed distance between the pair of series.
}
\description{
EDITH DISTANCE WITH REAL PENALTY 
Computes the Edit Distance with Real Penalty between a pair of numeric time series using the 'ERPDistance' function from the 'TSdist' package.
We by setting the 'sigma' parameter as the 10% of the time series length, and the 'g' parameter equal to 0.1.
}
\references{
for more details read the 'TSdist' package documentation.
}
