% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{EUC.DIST}
\alias{EUC.DIST}
\title{EUCLIDEAN DISTANCE WITH NA.OMIT}
\usage{
EUC.DIST(x, y)
}
\arguments{
\item{x}{a time series with real values.}

\item{y}{a time series with real values and not necessarily the same length as x.}
}
\value{
a positive real number result of the sum of point-to-point euclidean distance between x and y.
}
\description{
Compute point-to-point euclidean distance between two time series x and y.
}
