% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{Frechets.Measure}
\alias{Frechets.Measure}
\title{FRECHET DISTANCE}
\usage{
Frechets.Measure(s1, s2)
}
\arguments{
\item{s1}{a numeric vector containing the first time series.}

\item{s2}{a numeric vector containing the second time series.}
}
\value{
a positive real number resulting from the the calculated distance between the pair of series.
}
\description{
Computes the Frechet distance between two numerical trajectories.
}
