% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{GaussNoise}
\alias{GaussNoise}
\title{Adding White Gaussian Noise variation imposed}
\usage{
GaussNoise(S, sd = 1)
}
\arguments{
\item{S}{A time series with real values.}

\item{sd}{Standard deviation of the gaussian noise to be added.}
}
\value{
The input time series with the gaussian noise variation imposed.
}
\description{
The function is based on the add.Gaussian.noise() function from the RMThreshold package. Mean fixed in zero.
}
