% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_classifier_with_custom_distance_function.R
\name{knn}
\alias{knn}
\title{KNN prediction for several series}
\usage{
knn(x_fit, x_pred, k, func = "dist", weighted_pred = FALSE, p = NULL)
}
\arguments{
\item{x_fit}{a data frame with reference time series (by row) and first column as class label}

\item{x_pred}{a data frame with time series to classify (by row) and first column as class label}

\item{k}{number of nearest neighbors}

\item{func}{distance measure}

\item{weighted_pred}{whether to use weighted prediction or not}

\item{p}{extra parameter for the distance measure function}
}
\value{
a vector of predicted class labels
}
\description{
KNN prediction for several series
}
