% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_classifier_with_custom_distance_function.R
\name{knn_prediction}
\alias{knn_prediction}
\title{KNN Prediction Function}
\usage{
knn_prediction(x)
}
\arguments{
\item{x}{a vector of class labels}
}
\value{
the predicted class label
}
\description{
KNN Prediction Function
}
