% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{LCSS.Measure}
\alias{LCSS.Measure}
\title{LONGEST COMMON SUBSEQUENCE DISTANCE MEASURE}
\usage{
LCSS.Measure(s1, s2, e = 0.01)
}
\arguments{
\item{s1}{a numeric vector containing the first time series.}

\item{s2}{a numeric vector containing the second time series.}

\item{e}{a positive threshold value that defines the distance.}
}
\value{
a positive real number between 0 and 1. Where 0 is the maximum similarity and 1 the maximum dissimilarity.
}
\description{
Computes the 'Longest Common Sub sequence distance' between a pair of numeric time series, and return difference between the length of s1 and the LCSS distance.
}
\references{
for more details read the ''TSdist' package documentation
}
