% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_classifier_with_custom_distance_function.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{NEAREST NEIGHBORS}
\usage{
nearest_neighbors(x, obs, k, FUN, p = NULL)
}
\arguments{
\item{x}{an nxm matrix where each row is a reference time series. There are n series of length m.}

\item{obs}{a 1Xm matrix that represents a series of length m which we want to classify by comparing with the reference series.}

\item{k}{number of nearest neighbors}

\item{FUN}{distance measure.}

\item{p}{some extra parameter of the FUN function}
}
\description{
NEAREST NEIGHBORS
}
