% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{ReplARCont}
\alias{ReplARCont}
\title{Contaminación por otro proceso autorregresivo.}
\usage{
ReplARCont(y, prob, phi1)
}
\arguments{
\item{y}{The time serie as *numeric* to be contaminated.}

\item{prob}{The probability of success of the binomial distribution. A number between 0 and 1.}
}
\value{
The Y matrix with autoregressive process noise.
}
\description{
Se genrea un vector llamado "ts.cont" a partir de un vector inicial "y", y otrvector 
"ts.ar"de la misma dimensión. Los valores de "ts.ar"son generados con un proceso AR con parámetros φ. 
El prob,100% de los valores de "y"son elegidos al azar. Si y_i es seleccionado, entonces 
es reemplazado por la entrada i del vector "ts.ar". Este procedimiento es llamado 
contaminación de reemplazo por otro proceso AR al prob,100%.
}
