% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_indices.R
\name{SSIMT.NA.OMIT}
\alias{SSIMT.NA.OMIT}
\title{SSIMT INDEX WITH NA.OMIT}
\usage{
SSIMT.NA.OMIT(x, y)
}
\arguments{
\item{X}{a time series of length n with real values.}

\item{Y}{a time series of length n with real values.}
}
\value{
a real number between 1 and -1; where 1 means max similarity, -1 means opposite behavior and 0 no similarity.
}
\description{
Compute SSIMT similarity index between two same length time series X and Y.
}
