% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{TimeSca}
\alias{TimeSca}
\title{Time Scaling variation imposed}
\usage{
TimeSca(S, beta = 1)
}
\arguments{
\item{S}{A time series with real values.}

\item{beta}{a natural constant.}
}
\value{
The input time series with the time scaling variation imposed.
}
\description{
A time series distortion that modify the time length of a time series _S_.
}
\details{
$S1(t) = S(t*beta)$
}
