% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{TimeShift}
\alias{TimeShift}
\title{Time Shift variation imposed}
\usage{
TimeShift(S, beta = 1)
}
\arguments{
\item{S}{A time series with real values.}

\item{beta}{A natural constant.}
}
\value{
The input time series with the time shift variation imposed.
}
\description{
A time series distortion that shift the time in a time series _S_.
}
\details{
$S1(t) = S(t+beta)$
}
