% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{TS.SaltPepper}
\alias{TS.SaltPepper}
\title{Salt and pepper variation imposed}
\usage{
TS.SaltPepper(S, epsi)
}
\arguments{
\item{S}{A time series with real values.}

\item{epsi}{contamination percentage for the salt and pepper noise [0,1].}
}
\value{
The input time series with the gaussian noise variation imposed.
}
\description{
Similar to salt and pepper image noise but applying to time series. In this case, salt and pepper variations correspond to max(S) and and min(S) respectively.
}
