% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{TStudentNoise}
\alias{TStudentNoise}
\title{t-Student Noise}
\usage{
TStudentNoise(y, prob, dfr)
}
\arguments{
\item{prob}{The probability of success of the binomial distribution. A number between 0 and 1.}

\item{dfr}{The degrees of freedom of the t-student distribution.}

\item{Y}{The time serie as *numeric* to be contaminated.}
}
\value{
The Y matrix with t-Student noise.
}
\description{
A vector ts.student is created where every element of _ts.student_ follow a 
t(df) distribution. The _prop_% of the elements of the array _Y_ are 
selected randomly. If Y_i_ is selected then it is replaced by _Y_i_ + _ts.student_i_.
}
