% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{AddGaussianNoise}
\alias{AddGaussianNoise}
\title{Aditive Gaussian Noise}
\usage{
AddGaussianNoise(y, prob, mu, sigma)
}
\arguments{
\item{prob}{The probability of success of the binomial distribution. A number between 0 and 1.}

\item{Y}{The time serie (as numeric) to be contaminated.}

\item{\mu}{The mean of the normal distribution.}

\item{\sigma}{The standard deviation of the normal distribution. A real number.}
}
\value{
The y matrix with aditive noise.
}
\description{
Generates a matrix _C_ from an initial matrix _Y_ and a matrix \Nu of the same 
dimension. \Nu values follow an $N(\mu, \sigma^2)$ distribution. The _prop_% of 
the elements of the array _Y_ are selected randomly. If $Y_{i,j$} is selected 
then it is replaced by $Y_{i,j} + \Nu_{i,j}$. This procedure is called additive 
contamination of _prob_%.
}
