% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{AddGaussianNoise}
\alias{AddGaussianNoise}
\title{Aditive Gaussian Noise}
\usage{
AddGaussianNoise(y, prob, mu, sigma)
}
\arguments{
\item{y}{The time serie (as numeric) to be contaminated.}

\item{prob}{The probability of success of the binomial distribution. A number between 0 and 1.}

\item{mu}{The mean of the normal distribution.}

\item{sigma}{The standard deviation of the normal distribution. A real number.}
}
\value{
The y matrix with aditive noise.
}
\description{
Bla bla bla
}
