% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Distortions.R
\name{TStudentNoise}
\alias{TStudentNoise}
\title{t-Student Noise}
\usage{
TStudentNoise(y, prob, dfr)
}
\arguments{
\item{prob}{The probability of success of the binomial distribution. A number between 0 and 1.}

\item{dfr}{The degrees of freedom of the t-student distribution.}

\item{Y}{The time serie as *numeric* to be contaminated.}
}
\description{
Bla bla bla
}
